+++ Common/src/main/java/fuzs/puzzleslib/client/gui/screens/Screens.java
@@ -2,9 +2,13 @@ package fuzs.puzzleslib.client.gui.screens;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.components.AbstractWidget;
+import net.minecraft.client.gui.components.Widget;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.entity.ItemRenderer;
+
+import java.util.List;
 
 /**
  * a helper class for accessing encapsulated fields on a screen
@@ -33,6 +37,12 @@ public interface Screens {
      * @return item renderer
      */
     ItemRenderer getItemRenderer(Screen screen);
+
+    /**
+     * @param screen    screen instance
+     * @return          the renderable buttons (different from children, also there are narratable widgets)
+     */
+    List<Widget> getRenderableButtons(Screen screen);
 
     /**
      * @param screen container screen instance