+++ Fabric/src/main/java/fuzs/puzzleslib/mixin/client/accessor/MinecraftAccessor.java
@@ -1,0 +1,13 @@
+package fuzs.puzzleslib.mixin.client.accessor;
+
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.searchtree.SearchRegistry;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(Minecraft.class)
+public interface MinecraftAccessor {
+
+    @Accessor
+    SearchRegistry getSearchRegistry();
+}